fix errors

+++ src/snowflake/snowpark/modin/plugin/docstrings/series_utils.py
@@ -2198,7 +2198,7 @@ class CombinedDatetimelikeProperties:
         >>> tdelta_idx
         TimedeltaIndex(['0 days 00:00:01', '0 days 00:00:02', '0 days 00:00:03'], dtype='timedelta64[ns]', freq=None)
         >>> tdelta_idx.seconds
-        Index([1, 2, 3], dtype='int32')
+        Index([1, 2, 3], dtype='int64')
         """
 
     @property
@@ -2260,7 +2260,7 @@ class CombinedDatetimelikeProperties:
                         '0 days 00:00:00.000003'],
                        dtype='timedelta64[ns]', freq=None)
         >>> tdelta_idx.microseconds
-        Index([1, 2, 3], dtype='int32')
+        Index([1, 2, 3], dtype='int64')
         """
 
     @property
@@ -2292,7 +2292,7 @@ class CombinedDatetimelikeProperties:
                         '0 days 00:00:00.000000003'],
                        dtype='timedelta64[ns]', freq=None)
         >>> tdelta_idx.nanoseconds
-        Index([1, 2, 3], dtype='int32')
+        Index([1, 2, 3], dtype='int64')
         """
 
     @property
